Moved database adapter v1 into its own folder
