+++ app/assets/js/src/database/adapterV1/getDaysV1.ts
