+++ app/assets/js/src/database/adapterV1/getTemplatesV1.ts
