+++ app/assets/js/src/database/index.ts
@@ -1,4 +1,1 @@
-export { getDaysV1 } from './getDaysV1';
-export { getDayTasksV1 } from './getDayTasksV1';
-export { getTasksV1 } from './getTasksV1';
-export { getTemplatesV1 } from './getTemplatesV1';
+export { adapterV1 } from './adapterV1';
